+++ changelog.md
@@ -4,6 +4,9 @@
 
 ## v0.1.0-beta.4
 
+- Added user settings for ImGui scaling.
+- Added ImGui font API (`imgui.font`).
+- Updated default executor font.
 - Fixed crashes when garbage-collecting node/object userdata.
 - Fixed tiny ImGui on macOS Retina with improved scaling and fonts (forked matcool/gd-imgui-cocos).